feat(date, date-range): rewrite Date component using date-fns and remove uncontrolled support
Removes `moment` package and all uses of it in carbon as it is marked as legacy by the maintainers.
Rewrites `Date` component using `date-fns/fp` utils to handle parsing dates and formatting strings.
Removes support for using component as uncontrolled: `value` and `onChange` are now required props.
Consuming projects should update their `onChange` handlers to use the `formattedValue` to update the
value state, the `rawValue` should be used to run validations and for storing in the backend: the
`rawValue` reported in the event.target.value will either a vaild ISO string when the current input
value parses correctly to a date or as `null` when it cannot. Formats are no longer configurable in
the locale object passed to the `I18nProvider` and are now handled internally: all separators are
now valid input values if they can be parsed to a valid date and the the input value will be
reformatted to the current locale's preferred format `onBlur`. The `DateRange` component has been
updated to use the new implementation of `Date`.

BREAKING CHANGE: Date component is no longer class based so any previous extension of it will no
longer work. It also now only works as a controlled component and requires both `value` and
`onChange` props to work: the `onChange` handler should also use the `formattedValue` to set the
value state and `rawValue` for validation and storing in the backend. The `DateRange` can now only
be used as a controlled component, `value` and `onChange` are required and `formattedValue` should
be used to update the state.

fix #2996, fix #4458